KG/requirements.txt
pandas
openpyxl
neo4j
javalang
pyyaml

# Optional: For LLM-enhanced DAO analysis
boto3
openai
python-dotenv
requests
anthropic
